Refactor test execution and environment checks into helper functions for improved modularity and readability
